Code Styling

+++ 1. Bronze/Bronze 4/[B4] 25704 출석 이벤트.cpp
@@ -29,5 +29,5 @@ int main()
     cin >> n >> p;
 
     double result = SaledPrice(n, p);
-    cout << (int)result << '\n';
+    cout << result << '\n';
 }
